Add kill_worker and fix resulting bugs
kill_worker exposed some new faults in poolboy, where the maximum size
of the pool could end up smaller than it should be after a worker was
killed.

To help in debugging this, a 'status' command was added to poolboy to
look at its internal state and a dynamic_precondition function was used
to compare the expected state to the actual state.

Numerous problems were discovered by doing this, poolboy had issues
around when it changed states, like allocating the last non-overflow
worker would leave the pool in 'ready' until the next request came in,
instead of immediately changing the state.