WIP: re-write the poolboy_eqc as grouped commands
There is a periodic failure of `poolboy_eqc:prop_parallel/0`. When the
counter-example is checked it passes sometimes. This suggests we want to
pulse the property to discover the scheduler interleaving that causes
the issue. Before adding pulse it made sense to port the test to the
modern style for ease of working/adapting it. This commit is WIP, it
includes pulse, but the pulse test fails for as yet unexamined reasons.

The periodic failure is around the `kill_idle_worker` command. It checks
out a worker and checks it straight back in without modifying the statem
state, but there is a race window
where a parallel call to checkout fails (due to the pool being full
temporarily) since the state was not updated with the above checkout.