Improve poolboy_statem
Introduce a proxy for worker actions - this removes the need for the
self() hack in the parallel property.

Model how workers are queued, with this change the model can predict
what workers are "checked in" and thus it is possible to implement
kill_idle_worker in an atomic way.

Remove the killed field, no need to track those.

Added command_precondition_common to not have to check if #state.pid is
undefined everywhere